chor: adds support for font scaling
On both iOS and Android, the platform have options for configuring font scale for the visually impaired.
On Android font scale happens automatically, on iOS the app have to handle it.

This adds Label.accessibilityAdjustFontSize to enable font scaling on iOS.
Also adds classes to Page fitting the font scaling.
And an Observable for getting the value.