backported: performance improvements
CSS-helper classes are new added by override the cssClasses-property on the view.
This remove the the need to add a global loaded-event on the views to update them.
view._onCssStateChange() are now only called on root-views and on ns-modal views,
previously we called _onCssStateChange() and all loaded views. That caused too many
updates and noticable slow downs mainly on iOS.